+++ src/formatter/mod.rs
@@ -5,9 +5,8 @@ use crate::parser::{
     Directive, DirectiveNode, JournalCstNode, JournalFile, SingleLineComment, TransactionEntry,
     TransactionNode,
 };
-use core::fmt::Write;
 
-const TRANSACTION_ENTRY_VALUE_SPACING: usize = 2;
+const TRANSACTION_ENTRY_VALUE_SPACING: u8 = 2;
 
 #[derive(Default)]
 pub(crate) struct FormatContentOptions {
@@ -26,37 +25,46 @@ impl FormatContentOptions {
 }
 
 #[cfg(test)]
-fn format_content(nodes: &JournalFile) -> String {
+fn format_content(nodes: &JournalFile) -> Vec<u8> {
     format_content_with_options(nodes, &FormatContentOptions::default())
 }
 
 pub(crate) fn format_content_with_options(
     nodes: &JournalFile,
     opts: &FormatContentOptions,
-) -> String {
-    let mut formatted = String::with_capacity(opts.estimated_length);
+) -> Vec<u8> {
+    let mut buffer = Vec::with_capacity(opts.estimated_length);
+    format_nodes(nodes, &mut buffer);
+    buffer
+}
 
-    for node in nodes {
+fn format_nodes(nodes: &JournalFile, buffer: &mut Vec<u8>) {
+    #[cfg(any(test, feature = "tracing"))]
+    {
+        let span = tracing::span!(tracing::Level::TRACE, "format_nodes");
+        let _enter = span.enter();
+        tracing::trace!("nodes={:#?}", nodes);
+    }
+
+    for node in nodes.iter() {
         match node {
             JournalCstNode::SingleLineComment(SingleLineComment {
                 content,
                 prefix,
-                colno,
+                indent,
                 ..
             }) => {
-                _ = writeln!(
-                    formatted,
-                    "{}{}{}",
-                    " ".repeat(*colno - 1),
-                    *prefix as u8 as char,
-                    content
-                );
+                spaces::extend(buffer, *indent as usize);
+                buffer.push(*prefix as u8);
+                buffer.extend_from_slice(content.as_ref());
             }
             JournalCstNode::EmptyLine { .. } => {
-                formatted.push('\n');
+                buffer.push(b'\n');
             }
             JournalCstNode::MultilineComment { content, .. } => {
-                _ = writeln!(formatted, "comment\n{content}end comment");
+                buffer.extend_from_slice(b"comment\n");
+                buffer.extend_from_slice(content.as_ref());
+                buffer.extend_from_slice(b"end comment\n");
             }
             JournalCstNode::DirectivesGroup {
                 nodes,
@@ -71,39 +79,34 @@ pub(crate) fn format_content_with_options(
                             comment,
                             ..
                         }) => {
-                            _ = write!(formatted, "{name} {content}");
+                            buffer.extend_from_slice(name.as_ref());
+                            buffer.push(b' ');
+                            buffer.extend_from_slice(content.as_ref());
 
                             if let Some(comment) = comment {
-                                _ = write!(
-                                    formatted,
-                                    "{}{}{}",
-                                    " ".repeat(
-                                        2 + max_name_content_len
-                                            - name.chars().count()
-                                            - content.chars().count()
-                                    ),
-                                    comment.prefix as u8 as char,
-                                    comment.content
+                                spaces::extend(
+                                    buffer,
+                                    (2 + *max_name_content_len) as usize
+                                        - name.chars_count()
+                                        - content.chars_count(),
                                 );
+                                buffer.push(comment.prefix as u8);
+                                buffer.extend_from_slice(comment.content.as_ref());
                             }
-
-                            formatted.push('\n');
                         }
                         DirectiveNode::Subdirective(content) => {
-                            _ = writeln!(formatted, "  {content}");
+                            spaces::extend(buffer, 2);
+                            buffer.extend_from_slice(content.as_ref());
                         }
                         DirectiveNode::SingleLineComment(SingleLineComment {
                             content,
                             prefix,
                             ..
                         }) => {
-                            _ = writeln!(
-                                formatted,
-                                "{}{}{}",
-                                " ".repeat(*max_name_content_len + 3),
-                                *prefix as u8 as char,
-                                content
-                            );
+                            spaces::extend(buffer, (*max_name_content_len + 3) as usize);
+                            buffer.push(*prefix as u8);
+                            buffer.extend_from_slice(content.as_ref());
+                            buffer.push(b'\n');
                         }
                     }
                 }
@@ -123,15 +126,13 @@ pub(crate) fn format_content_with_options(
                 max_entry_value_third_part_before_decimals_len,
                 max_entry_value_third_part_after_decimals_len,
             } => {
-                _ = write!(formatted, "{}", title.trim());
+                buffer.extend_from_slice(title.as_ref());
                 if let Some(comment) = title_comment {
-                    _ = write!(
-                        formatted,
-                        "  {}{}",
-                        comment.prefix as u8 as char, comment.content
-                    );
+                    spaces::extend(buffer, 2);
+                    buffer.push(comment.prefix as u8);
+                    buffer.extend_from_slice(comment.content.as_ref());
                 }
-                formatted.push('\n');
+                buffer.push(b'\n');
 
                 for entry in entries {
                     match entry {
@@ -149,134 +150,164 @@ pub(crate) fn format_content_with_options(
                                 comment,
                             } = inner.as_ref();
 
-                            let entry_line = format!(
-                                "{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}",
-                                " ".repeat(*first_entry_indent),
-                                name,
-                                if !value_first_part_before_decimals.is_empty() {
-                                    let name_len = name.chars().count();
-                                    let before_decimals_len =
-                                        value_first_part_before_decimals.chars().count();
-                                    let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
-                                        + max_entry_name_len
-                                        - name_len
-                                        + max_entry_value_first_part_before_decimals_len
-                                        - before_decimals_len;
-                                    " ".repeat(n_spaces)
-                                } else {
-                                    "".to_string()
-                                },
-                                value_first_part_before_decimals,
-                                value_first_part_after_decimals,
-                                if !value_first_separator.is_empty() {
-                                    let after_decimals_len =
-                                        value_first_part_after_decimals.chars().count();
-                                    let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
-                                        + max_entry_value_first_part_after_decimals_len
-                                        - after_decimals_len;
-                                    " ".repeat(n_spaces)
-                                } else {
-                                    "".to_string()
-                                },
-                                value_first_separator,
-                                if !value_second_part_before_decimals.is_empty() {
-                                    let value_first_separator_len = value_first_separator.len();
-                                    let value_second_part_before_decimals_len =
-                                        value_second_part_before_decimals.chars().count();
-                                    let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
-                                        + max_entry_value_first_separator_len
-                                        - value_first_separator_len
-                                        + max_entry_value_second_part_before_decimals_len
-                                        - value_second_part_before_decimals_len;
-                                    " ".repeat(n_spaces)
-                                } else {
-                                    "".to_string()
-                                },
-                                value_second_part_before_decimals,
-                                value_second_part_after_decimals,
-                                if !value_second_separator.is_empty() {
-                                    let value_second_part_after_decimals_len =
-                                        value_second_part_after_decimals.chars().count();
-                                    let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
-                                        + max_entry_value_second_part_after_decimals_len
-                                        - value_second_part_after_decimals_len;
-                                    " ".repeat(n_spaces)
-                                } else {
-                                    "".to_string()
-                                },
-                                value_second_separator,
-                                if !value_third_part_before_decimals.is_empty() {
-                                    let value_second_separator_len = value_second_separator.len();
-                                    let value_third_part_before_decimals_len =
-                                        value_third_part_before_decimals.chars().count();
-                                    let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
-                                        + max_entry_value_second_separator_len
-                                        - value_second_separator_len
-                                        + max_entry_value_third_part_before_decimals_len
-                                        - value_third_part_before_decimals_len;
-                                    " ".repeat(n_spaces)
-                                } else {
-                                    "".to_string()
-                                },
-                                value_third_part_before_decimals,
-                                value_third_part_after_decimals,
+                            let mut entry_line_buffer = Vec::new();
+
+                            spaces::extend(&mut entry_line_buffer, *first_entry_indent as usize);
+                            entry_line_buffer.extend_from_slice(name.as_ref());
+                            if !value_first_part_before_decimals.is_empty() {
+                                let name_len = name.chars_count() as u8;
+                                let before_decimals_len =
+                                    value_first_part_before_decimals.chars_count() as u8;
+                                let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
+                                    + max_entry_name_len
+                                    - name_len
+                                    + max_entry_value_first_part_before_decimals_len
+                                    - before_decimals_len;
+                                spaces::extend(&mut entry_line_buffer, n_spaces as usize);
+                            }
+                            entry_line_buffer.extend_from_slice(
+                                value_first_part_before_decimals.as_ref(),
+                            );
+                            entry_line_buffer.extend_from_slice(
+                                value_first_part_after_decimals.as_ref(),
                             );
 
-                            formatted.push_str(&entry_line);
+                            if !value_first_separator.is_empty() {
+                                let after_decimals_len =
+                                    value_first_part_after_decimals.chars_count() as u8;
+                                let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
+                                    + max_entry_value_first_part_after_decimals_len
+                                    - after_decimals_len;
+                                spaces::extend(&mut entry_line_buffer, n_spaces as usize);
+                            }
+                            entry_line_buffer.extend_from_slice(value_first_separator.as_ref());
+                            if !value_second_part_before_decimals.is_empty() {
+                                let value_first_separator_len =
+                                    value_first_separator.len() as u8;
+                                let value_second_part_before_decimals_len =
+                                    value_second_part_before_decimals.chars_count() as u8;
+                                let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
+                                    + max_entry_value_first_separator_len
+                                    - value_first_separator_len
+                                    + max_entry_value_second_part_before_decimals_len
+                                    - value_second_part_before_decimals_len;
+                                spaces::extend(&mut entry_line_buffer, n_spaces as usize);
+                            }
+                            entry_line_buffer.extend_from_slice(
+                                value_second_part_before_decimals.as_ref(),
+                            );
+                            entry_line_buffer.extend_from_slice(
+                                value_second_part_after_decimals.as_ref(),
+                            );
+
+                            if !value_second_separator.is_empty() {
+                                let value_second_part_after_decimals_len =
+                                    value_second_part_after_decimals.chars_count() as u8;
+                                let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
+                                    + max_entry_value_second_part_after_decimals_len
+                                    - value_second_part_after_decimals_len;
+                                spaces::extend(&mut entry_line_buffer, n_spaces as usize);
+                            }
+                            entry_line_buffer.extend_from_slice(value_second_separator.as_ref());
+                            if !value_third_part_before_decimals.is_empty() {
+                                let value_second_separator_len =
+                                    value_second_separator.len() as u8;
+                                let value_third_part_before_decimals_len =
+                                    value_third_part_before_decimals.chars_count() as u8;
+                                let n_spaces = TRANSACTION_ENTRY_VALUE_SPACING
+                                    + max_entry_value_second_separator_len
+                                    - value_second_separator_len
+                                    + max_entry_value_third_part_before_decimals_len
+                                    - value_third_part_before_decimals_len;
+                                spaces::extend(&mut entry_line_buffer, n_spaces as usize);
+                            }
+                            entry_line_buffer.extend_from_slice(
+                                value_third_part_before_decimals.as_ref(),
+                            );
+                            entry_line_buffer.extend_from_slice(
+                                value_third_part_after_decimals.as_ref(),
+                            );
+
+
 
                             if let Some(comment) = comment {
                                 let comment_separation = if !value_second_separator.is_empty() {
                                     TRANSACTION_ENTRY_VALUE_SPACING
                                         + max_entry_value_third_part_after_decimals_len
-                                        - value_third_part_after_decimals.chars().count()
+                                        - value_third_part_after_decimals.chars_count() as u8
                                 } else if !value_first_separator.is_empty() {
                                     TRANSACTION_ENTRY_VALUE_SPACING
                                         + max_entry_value_second_part_after_decimals_len
-                                        - value_second_part_after_decimals.chars().count()
+                                        - value_second_part_after_decimals.chars_count() as u8
                                 } else {
                                     TRANSACTION_ENTRY_VALUE_SPACING
                                         + max_entry_value_first_part_after_decimals_len
-                                        - value_first_part_after_decimals.chars().count()
+                                        - value_first_part_after_decimals.chars_count() as u8
                                 };
 
-                                let title_chars_count = title.chars().count();
-                                let entry_line_chars_count = entry_line.chars().count();
+                                let entry_line_chars_count = String::from_utf8_lossy(&entry_line_buffer).chars().count();
 
-                                _ = write!(
-                                    formatted,
-                                    "{}{}{}",
-                                    " ".repeat(
-                                        if title_chars_count + 2 > entry_line_chars_count + 2 {
-                                            title_chars_count + 2 - entry_line_chars_count
-                                        } else {
-                                            comment_separation
-                                        }
-                                    ),
-                                    comment.prefix as u8 as char,
-                                    comment.content
-                                );
-                            };
+                                buffer.append(&mut entry_line_buffer);
 
-                            formatted.push('\n');
+                                let title_chars_count = title.chars_count();
+
+                                let n_spaces = if title_chars_count + 2 > entry_line_chars_count + 2
+                                {
+                                    title_chars_count + 2 - entry_line_chars_count
+                                } else {
+                                    comment_separation as usize
+                                };
+                                spaces::extend(buffer, n_spaces);
+                                buffer.push(comment.prefix as u8);
+                                buffer.extend_from_slice(comment.content.as_ref());
+                            } else {
+                                buffer.append(&mut entry_line_buffer);
+                            }
+
+                            buffer.push(b'\n');
                         }
                         TransactionNode::SingleLineComment(SingleLineComment {
                             content,
                             prefix,
                             ..
                         }) => {
-                            _ = writeln!(
-                                formatted,
-                                "{}{}{}",
-                                " ".repeat(*first_entry_indent),
-                                *prefix as u8 as char,
-                                content
-                            );
+                            spaces::extend(buffer, *first_entry_indent as usize);
+                            buffer.push(*prefix as u8);
+                            buffer.extend_from_slice(content.as_ref());
+                            buffer.push(b'\n');
                         }
                     }
                 }
             }
         }
+
+
+
+        #[cfg(any(test, feature = "tracing"))]
+        {
+            let span = tracing::span!(tracing::Level::TRACE, "format_nodes(formatted)");
+            let _enter = span.enter();
+
+            tracing::trace!("buffer=\"{}\"", String::from_utf8_lossy(buffer));
+
+        }
+    }
+}
+
+mod spaces {
+    /// Generate a compile-time array of spaces of size N,
+    /// to avoid allocating strings at runtime using " ".repeat(N).
+    const fn make_spaces<const N: usize>() -> [u8; N] {
+        [b' '; N]
     }
 
-    formatted
+    const SPACES_64: [u8; 64] = make_spaces::<64>();
+
+    pub fn extend(buffer: &mut Vec<u8>, n: usize) {
+        if n <= 64 {
+            buffer.extend_from_slice(&SPACES_64[..n]);
+        } else {
+            buffer.extend(std::iter::repeat(b' ').take(n));
+        }
+    }
 }